Move EventQuery to its own file
EventQuery struct and its CustomStringConvertible extension were relocated from Event.swift to a new EventQuery.swift file for better code organization and separation of concerns.
